A working APK with a one sided chatbox. This is just a version of chat as provided by param bhaiya. Multi User chatbox with notifications will be updated in a few days.
