+++ app/build.gradle
@@ -36,10 +36,15 @@ dependencies {
     implementation 'com.android.support.constraint:constraint-layout:1.1.3'
     implementation 'com.firebase:firebase-client-android:2.3.1'
     implementation 'com.android.support:design:28.0.0'
+    implementation platform('com.google.firebase:firebase-bom:17.0.0')
     implementation 'com.google.firebase:firebase-crash:16.2.1'
     implementation 'com.google.firebase:firebase-storage:17.0.0'
     implementation 'com.github.tibolte:agendacalendarview:1.0.4'
-    implementation 'com.android.support:support-v4:28.0.0-rc02'
+    implementation 'com.firebaseui:firebase-ui-database:4.2.0'
+    implementation 'de.hdodenhof:circleimageview:3.0.0'
+    implementation 'com.github.bumptech.glide:glide:4.9.0'
+    annotationProcessor 'com.github.bumptech.glide:compiler:4.9.0'
+    implementation 'com.android.support:support-v4:28.0.0'
     implementation 'com.android.support:support-vector-drawable:28.0.0'
     implementation "com.google.firebase:firebase-messaging:$Firebase_messaging_version"
     testImplementation 'junit:junit:4.12'
