+++ app/src/main/AndroidManifest.xml
@@ -14,17 +14,18 @@
         android:supportsRtl="true"
         android:theme="@style/AppTheme"
         tools:ignore="GoogleAppIndexingWarning">
-        <service android:name="com.example.apple.myapplication.MyFirebaseMessagingService">
+        <activity android:name=".ChatActivityDefault"></activity>
+
+        <service android:name=".MyFirebaseMessagingService">
             <intent-filter>
                 <action android:name="com.google.firebase.MESSAGING_EVENT" />
             </intent-filter>
-        </service>
-        <!--service-->
-            <!--android:name=".MyFcmListenerService">-->
-            <!--<intent-filter>-->
-                <!--<action android:name="com.google.firebase.MESSAGING_EVENT" />-->
-            <!--</intent-filter>-->
-        <!--</service>-->
+        </service> <!-- service -->
+        <!-- android:name=".MyFcmListenerService"> -->
+        <!-- <intent-filter> -->
+        <!-- <action android:name="com.google.firebase.MESSAGING_EVENT" /> -->
+        <!-- </intent-filter> -->
+        <!-- </service> -->
         <activity
             android:name=".MainActivity"
             android:label="@string/app_name"
@@ -51,9 +52,7 @@
             android:name=".CourseInfoActivity"
             android:label="Course Info" />
         <activity android:name=".ProfessorActivity" />
-        <activity android:name=".ClubActivity"></activity>
-
-
+        <activity android:name=".ClubActivity" />
     </application>
 
 </manifest>
