+++ app/src/main/java/com/example/apple/myapplication/Preferences.java
@@ -1,6 +1,6 @@
 package com.example.apple.myapplication;
 
-public class Prefrences {
+public class Preferences {
     public static final String FRIENDLY_MSG_LENGTH = "friendly_msg_length";
 
 }
