--- app/src/main/java/com/example/apple/myapplication/SnapshotParser.java
@@ -1,4 +1,0 @@
-package com.example.apple.myapplication;
-
-class SnapshotParser<T> {
-}
