+++ app/src/main/res/menu/popup_menu_mailbox.xml
@@ -13,4 +13,8 @@
         android:id="@+id/three"
         android:title="Lost and Found Box" />
 
+    <item
+        android:id="@+id/four"
+        android:title="Chat Box" />
+
 </menu>
